refactor: update payment processing to use Decimal for monetary values
- Updated AbstractPaymentService to accept Decimal instead of float for payment processing.
- Modified PaymentService to handle Decimal values in payment attempts.
- Adjusted order and product schemas to use Decimal for prices and total amounts.
- Refactored database models to store prices as Numeric(10, 2) for precision.
- Enhanced order creation logic to compute total prices using Decimal.
- Implemented pagination for order and product listing endpoints.
- Added ListOrders use case to handle order retrieval based on user roles.
- Updated tests to reflect changes in data types and pagination structure.

+++ src/core/entities/order_item.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass
+from decimal import Decimal
 from uuid import UUID
 
 
@@ -7,4 +8,4 @@ class OrderItem:
     id: UUID
     name: str
     variation: str
-    unit_price: float  # base_price + price_change
+    unit_price: Decimal
