+++ src/core/services/abstract_payment_service.py
@@ -1,7 +1,8 @@
 from abc import ABC, abstractmethod
+from decimal import Decimal
 
 
 class AbstractPaymentService(ABC):
     @abstractmethod
-    async def process(self, value: float) -> dict:
+    async def process(self, value: Decimal) -> dict:
         ...
